+++ android/src/main/java/map/naver/plugin/net/lbstech/naver_map_plugin/NaverMapOptionSink.java
@@ -21,6 +21,8 @@ public interface NaverMapOptionSink {
 
     void setLocationButtonEnable(boolean locationButtonEnable);
 
+    void setContentPadding(List<Double> paddingData);
+
     /**
      * flutter 에서 setState()로 값을 변경해도 반영되지 않는 method. 최초 생성시에만 값변경.
      */