+++ lib/src/marker.dart
@@ -9,17 +9,17 @@ class Marker {
 
   /// 마커를 클릭했을 때 인포윈도우를 보이게 할 수 있습니다.
   /// null 일 경우 마커를 클릭해도 윈도우가 보이지 않습니다.
-  final String infoWindow;
+  String infoWindow;
 
   /// 마커의 불투명도를 조절하는 값입니다.
   /// 유효한 값의 범위는 0.0 ~ 1.0 까지이며 1.0일떄 완전 붏투명입니다.
-  final double alpha;
+  double alpha;
 
   /// 마커를 평평하게 설정할지 여부를 지정합니다.
   ///
   /// 마커가 평평할 경우 지도가 회전하거나 기울어지면 마커 이미지도 함께 회전하거나
   /// 기울어집니다. 단, 마커가 평평하더라도 이미지의 크기는 항상 동일하게 유지됩니다.
-  final bool flat;
+  bool flat;
 
   /// 오직 클릭 이벤트 리스너가 지정된 오버레이만이 클릭 이벤트를 받을 수 있습니다.
   ///
@@ -40,36 +40,36 @@ class Marker {
   /// 캡션의 텍스트를 지정합니다. 빈 문자열일 경우 캡션이 그려지지 않습니다.
   /// -
   /// 기본값은 빈 문자열입니다.
-  final String captionText;
+  String captionText;
 
   /// 캡션의 크기를 지정합니다.
-  final double captionTextSize;
+  double captionTextSize;
 
   /// 캡션이 보이는 최대 줌 레벨을 지정합니다.
   ///
   ///
   /// 지도의 줌 레벨이 캡션의 최대 줌 레벨보다 클 경우 아이콘만 나타나고
   /// 캡션은 화면에 나타나지 않으며 이벤트도 받지 못합니다.
-  final double captionMaxZoom;
+  double captionMaxZoom;
 
   /// 캡션이 보이는 최소 줌 레벨을 지정합니다.
   ///
   ///
   /// 지도의 줌 레벨이 캡션의 최소 줌 레벨보다 작을 경우 아이콘만 나타나고
   /// 캡션은 화면에 나타나지 않으며 이벤트도 받지 못합니다.
-  final double captionMinZoom;
+  double captionMinZoom;
 
   /// 캡션의 색상을 지정합니다. RGB CODE를 필요로 합니다.
   ///
   ///
   /// 기본값은 [Colors.black]입니다.
-  final Color captionColor;
+  Color captionColor;
 
   /// 캡션의 테두리색을 지정합니다. RGB Code를 필요로합니다.
   ///
   ///
   /// 기본값은 [Colors.white] 입니다.
-  final Color captionHaloColor;
+  Color captionHaloColor;
 
   /// 캡션의 희망 너비를 지정합니다.
   ///
@@ -79,22 +79,22 @@ class Marker {
   ///
   /// 하나의 어절이 길 경우 캡션의 너비가 희망 너비를 초과할 수 있습니다.
   /// 0일 경우 너비를 제한하지 않습니다.
-  final int captionRequestedWidth;
+  int captionRequestedWidth;
 
   /// 아이콘과 캡션 간의 여백을 지정합니다.
   /// </br>
   /// 단위 >> Android: DP, iOS: PT, 기본값은 0입니다.
-  final int captionOffset;
+  int captionOffset;
 
   /// 보조 캡션의 텍스트를 지정합니다. 보조 캡션은 주 캡션의 하단에 나타납니다.
   /// 빈 문자열일 경우 보조 캡션이 그려지지 않습니다.
   ///
   ///
   /// 기본값은 빈 문자열입니다.
-  final String subCaptionText;
+  String subCaptionText;
 
   /// 보조 캡션의 텍스트 크기를 지정합니다.
-  final double subCaptionTextSize;
+  double subCaptionTextSize;
 
   /// 보조 캡션의 너비를 지정합니다. 지정할 경우 한 줄의 너비가 희망 너비를 초과하는
   /// 캡션 텍스트는 자동으로 개행됩니다. 자동 개행은 어절 단위로 이루어지므로,
@@ -102,57 +102,57 @@ class Marker {
   ///
   ///
   /// 0일 경우 너비를 제한하지 않습니다.
-  final int subCaptionRequestedWidth;
+  int subCaptionRequestedWidth;
 
   /// 보조 캡션의 색상을 지정합니다.
   ///
   ///
   /// 기본값은 [Colors.black]입니다.
-  final Color subCaptionColor;
+  Color subCaptionColor;
 
   /// 보조 캡션의 테두리 색상을 지정합니다.
   ///
   ///
   /// 기본값은 [Colors.white]입니다.
-  final Color subCaptionHaloColor;
+  Color subCaptionHaloColor;
 
   /// 아이콘의 너비, 높이를 지정합니다.
   /// <br/>
   /// 단위 >> android : dp, iOS: pt
   /// <br/>
   /// 값이 없는 경우 이미지의 너비를 따릅니다.
-  final int width, height;
+  int width, height;
 
   /// 오버레이가 보이는 최대, 최소 줌 레벨을 지정합니다. 지도의 줌 레벨이
   /// 오버레이의 최대 줌 레벨보다 크거나 지도의 줌 레벨이 오보레이의
   /// 줌 레벨보다 작은 경우 오버레이는 화면에 나타나지 않으며
   /// 이벤트도 받지 못합니다.
-  final double maxZoom, minZoom;
+  double maxZoom, minZoom;
 
   /// 아이콘의 각도를 지정합니다. 각도를 지정하면 아이콘이 해당 각도만큼
   /// 시계 방향으로 회전합니다
-  final double angle;
+  double angle;
 
   /// 아이콘을 지정합니다.
-  final OverlayImage icon;
+  OverlayImage icon;
 
   /// 캡션에 원근 효과를 적용할지 여부를 반환합니다.
   /// 원근 효과를 적용할 경우 가까운 캡션은 크게, 먼 캡션은 작게 표시됩니다.
   ///
   ///
   /// 기본값은 false입니다.
-  final bool captionPerspectiveEnabled;
+  bool captionPerspectiveEnabled;
 
   /// 아이콘에 덧입힐 색상을 지정합니다. 덧입힐 색상을 지정하면 덧입힐 색상이
   /// 아이콘 이미지의 색상과 가산 혼합됩니다. 단, 덧입힐 색상의 알파는 무시됩니다.
   ///
   ///
   /// 기본값은 Color.TRANSPARENT입니다.
-  final Color iconTintColor;
+  Color iconTintColor;
 
   /// 보조 Z 인덱스를 지정합니다. 전역 Z 인덱스가 동일한 여러 오버레이가 화면에서
   /// 겹쳐지면 보조 Z 인덱스가 큰 오버레이가 작은 오버레이를 덮습니다.
-  final int zIndex;
+  int zIndex;
 
   /// 전역 Z 인덱스를 지정합니다. 여러 오버레이가 화면에서 겹쳐지면 전역 Z
   /// 인덱스가 큰 오버레이가 작은 오버레이를 덮습니다.
@@ -161,14 +161,14 @@ class Marker {
   ///
   ///
   /// 기본값은 DEFAULT_GLOBAL_Z_INDEX입니다
-  final int globalZIndex;
+  int globalZIndex;
 
   /// ## 마커 아이콘의 앵커를 지정합니다.
   /// 앵커로 지정된 지점이 정보창의 좌표에 위치합니다.
   /// 값의 범위는 x, y 각각 0 ~ 1 이며, (0,0)일 경우 좌상단, (1,1)일 경우 우츨 하단을 의미합니다.
   /// 기본값은 중앙 하단인 (0.5, 1)입니다.
   /// > 앵커는 아이콘 이미지에서 기준이 되는 지점을 의미하는 값으로, 아이콘에서 앵커로 지정된 지점이 마커의 좌표에 위치하게 됩니다.
-  final AnchorPoint anchor;
+  AnchorPoint anchor;
 
   Marker(
       {@required this.markerId,